Organize store_app/common
It was really hard to keep an eye on the content inside store_app/common
This organizes general purpose code in common/, snap related code in snap/, and packagekit related code in packagekit/
