+++ lib/store_app/common/app_page/app_description.dart
