+++ lib/store_app/common/snap/snap_model.dart
