+++ lib/store_app/common/snap/snap_section.dart
