+++ lib/store_app/explore/explore_page.dart
@@ -21,10 +21,10 @@ import 'package:software/l10n/l10n.dart';
 import 'package:software/services/package_service.dart';
 import 'package:software/services/snap_service.dart';
 import 'package:software/store_app/common/constants.dart';
-import 'package:software/store_app/common/offline_page.dart';
-import 'package:software/store_app/common/package_page.dart';
-import 'package:software/store_app/common/snap_page.dart';
-import 'package:software/store_app/common/snap_section.dart';
+import 'package:software/store_app/explore/offline_page.dart';
+import 'package:software/store_app/common/packagekit/package_page.dart';
+import 'package:software/store_app/common/snap/snap_page.dart';
+import 'package:software/store_app/common/snap/snap_section.dart';
 import 'package:software/store_app/explore/explore_model.dart';
 import 'package:software/store_app/explore/search_field.dart';
 import 'package:software/store_app/explore/search_page.dart';
