+++ lib/store_app/explore/offline_page.dart
