+++ lib/store_app/explore/search_page.dart
@@ -26,9 +26,9 @@ import 'package:software/store_app/common/app_format.dart';
 import 'package:software/store_app/common/app_format_popup.dart';
 import 'package:software/store_app/common/app_icon.dart';
 import 'package:software/store_app/common/constants.dart';
-import 'package:software/store_app/common/packagekit_filter_button.dart';
-import 'package:software/store_app/common/packagekit_group_button.dart';
-import 'package:software/store_app/common/snap_section_popup.dart';
+import 'package:software/store_app/common/packagekit/packagekit_filter_button.dart';
+import 'package:software/store_app/common/packagekit/packagekit_group_button.dart';
+import 'package:software/store_app/common/snap/snap_section_popup.dart';
 import 'package:software/store_app/explore/explore_model.dart';
 import 'package:yaru_icons/yaru_icons.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
