+++ lib/store_app/explore/start_page.dart
@@ -19,7 +19,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:snapd/snapd.dart';
 import 'package:software/l10n/l10n.dart';
-import 'package:software/store_app/common/snap_section.dart';
+import 'package:software/store_app/common/snap/snap_section.dart';
 import 'package:software/store_app/explore/color_banner.dart';
 import 'package:software/store_app/explore/explore_model.dart';
