+++ lib/store_app/my_apps/my_apps_page.dart
@@ -22,8 +22,8 @@ import 'package:software/services/package_service.dart';
 import 'package:software/services/snap_service.dart';
 import 'package:software/store_app/common/app_format.dart';
 import 'package:software/store_app/common/app_format_popup.dart';
-import 'package:software/store_app/common/package_page.dart';
-import 'package:software/store_app/common/snap_page.dart';
+import 'package:software/store_app/common/packagekit/package_page.dart';
+import 'package:software/store_app/common/snap/snap_page.dart';
 import 'package:software/store_app/my_apps/my_apps_model.dart';
 import 'package:software/store_app/my_apps/my_apps_search_field.dart';
 import 'package:software/store_app/my_apps/my_packages_page.dart';
